+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * added `task` service
 * added Mysql GIS and Oracle GIS dj-database-url alias
 * added MSSQL, Redshift, CockroachDB and Timescale service urls
+* removed `Service.validate()` method
 
 ## 1.9.0